+++ carp_mobile_sensing/lib/probes/connectivity/bluetooth_probe.dart
@@ -10,6 +10,8 @@ part of connectivity;
 // This probe requests access to location PERMISSIONS (on Android). Don't ask why.....
 // TODO - implement request for getting permission.
 
+// TODO - need to reimplement this probe -- taking into consideration if BT is available and only start scans when asked...
+
 /// The [BluetoothProbe] scans for nearby and visible Bluetooth devices and collect
 /// a [BluetoothDatum] for each. Uses a [PeriodicMeasure] for configuration the
 /// [frequency] and [duration] of the scan.
@@ -17,13 +19,18 @@ class BluetoothProbe extends PeriodicStreamProbe {
   /// Default timeout for bluetooth scan - 2 secs
   static const DEFAULT_TIMEOUT = 2 * 1000;
 
-  BluetoothProbe(PeriodicMeasure measure)
-      : super(
-            measure,
-            FlutterBlue.instance
-                .scan(
-                    scanMode: ScanMode.lowLatency, timeout: Duration(milliseconds: measure.duration ?? DEFAULT_TIMEOUT))
-                .map((result) => BluetoothDatum.fromScanResult(result)));
+//  BluetoothProbe(PeriodicMeasure measure)
+//      : super(
+//            measure,
+//            FlutterBlue.instance
+//                .scan(
+//                    scanMode: ScanMode.lowLatency, timeout: Duration(milliseconds: measure.duration ?? DEFAULT_TIMEOUT))
+//                .map((result) => BluetoothDatum.fromScanResult(result)));
+
+  BluetoothProbe()
+      : super(FlutterBlue.instance
+            .scan(scanMode: ScanMode.lowLatency, timeout: Duration(milliseconds: DEFAULT_TIMEOUT))
+            .map((result) => BluetoothDatum.fromScanResult(result)));
 
   // important that we don't propagate this up the chain, since this would close the overall stream
   @override